requirements.txt
allure-pytest==2.9.42
allure-python-commons==2.9.42
appdirs==1.4.4
atomicwrites==1.4.0
attrs==21.2.0
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
configparser==5.0.2
crayons==0.4.0
distlib==0.3.2
et-xmlfile==1.1.0
filelock==3.0.12
idna==2.10
iniconfig==1.1.1
jdcal==1.4.1
openpyexcel==2.5.14
packaging==20.9
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
requests==2.25.1
selenium==3.141.0
six==1.16.0
toml==0.10.2
urllib3==1.26.4
virtualenv==20.4.7
webdriver-manager==3.4.2
